pyproject.toml
[tool.poetry]
name = "metom"
version = "0.1.0"
description = "MLOps for End-to-end Training of Online Models"
authors = ["Haldun Bucak, Ilkin Berker Topuz"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.5.1"
torchvision = "^0.20.1"
pandas = "^2.2.3"
pyyaml = "^6.0.2"
scikit-learn = "^1.5.2"
numpy = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
